Address Proton profiler review feedback
Limit the initial integration to NVIDIA/CUPTI, skip annotation work outside active profiling windows, and consolidate Proton coverage into the existing GPU profiler suite with behavior-based recovery tests.\n\nAssisted-by: OpenAI Codex
