+++ vllm/config/profiler.py
@@ -14,7 +14,7 @@ from vllm.utils.hashing import safe_hash
 logger = init_logger(__name__)
 
 ProfilerKind = Literal["torch", "cuda", "proton"]
-ProtonBackend = Literal["cupti", "rocprofiler"]
+ProtonBackend = Literal["cupti"]
 ProtonContext = Literal["shadow", "python"]
 ProtonData = Literal["tree", "trace"]
 ProtonHook = Literal["triton"]
@@ -64,7 +64,7 @@ class ProfilerConfig:
     produces a Chrome trace."""
 
     proton_backend: ProtonBackend | None = None
-    """Proton GPU backend. ``None`` lets Proton select the platform backend."""
+    """Proton GPU backend. ``None`` lets Proton select CUPTI automatically."""
 
     proton_mode: str | None = None
     """Optional backend-specific Proton mode string, such as ``pcsampling``."""
